+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     Suitable for migration tests.
 * [dropwizard]
   - Add --dry-run option for update simulation. Prints merged config to console.
+* [cli]
+  - Add --dry-run option for update simulation. Prints merged config to console.
 
 ### 1.1.0 (2021-08-31)
 * [api]